Improved the zipfile build script; add description of CPX; Bump up version to 1.3.3

+++ make_zip.sh
@@ -4,6 +4,13 @@
 
 CURRENT_PATH=$(pwd)
 
+if [ ! -f docs/conf.py ]; then
+    echo "Could not find Sphinx configuration file at ./docs/conf.py"
+    exit 1
+fi
+RELEASE=`cd docs; python3 -c "import conf; print(conf.release)"`
+echo "RELEASE is $RELEASE"
+
 TMP_PATH=$(mktemp -d)
 trap "rm -rf $TMP_PATH" EXIT  # Ensure temporary path gets removed on exit
 cd $TMP_PATH
@@ -30,35 +37,52 @@ fi
 
 set -e  # Stop if there is any error
 
+# create root dir
+mkdir micropython-iot-software
+
 # Get tool wheels
 $PIP wheel -w micropython-iot-software/python-tools --no-binary=:all: esptool adafruit-ampy \
     wheel paho_mqtt thingflow git+https://github.com/wendlers/mpfshell
 
 # Get firmware
 cd micropython-iot-software
+BASE=`pwd` # save for later
 wget http://micropython.org/resources/firmware/esp8266-20170612-v1.9.1.bin
 
 # Get repos
-git clone https://github.com/jfischer/micropython-iot-hackathon
 git clone https://github.com/mpi-sws-rse/thingflow-python
-rm -rf micropython-iot-hackathon/.git
 rm -rf thingflow-python/.git
+git clone https://github.com/jfischer/micropython-iot-hackathon
+mv micropython-iot-hackathon/example_code ./example_code
 
 # build the docs
 echo "building docs"
 pip install sphinx_rtd_theme # theme for docs
 mkdir docs
+mv micropython-iot-hackathon/lecture.pdf ./docs
 cd micropython-iot-hackathon/docs
 make html
-mv _build/html ../../docs/micropython-iot-hackathon
-rm -rf _build
-cd ../..
+mv _build/html $BASE/docs/micropython-iot-hackathon
+cd $BASE
+rm -rf micropython-iot-hackathon # no longer need the original code
 cd thingflow-python/docs
 make html
-mv _build/html ../../docs/thingflow-python
+mv _build/html $BASE/docs/thingflow-python
 rm -rf _build
-cd ../..
+cd $BASE
 
+# clone and build micropython docs
+mkdir $BASE/mpbuild
+cd $BASE/mpbuild
+git clone https://github.com/micropython/micropython.git
+cd micropython
+git checkout -b v1.9.1
+cd docs
+make MICROPY_PORT=esp8266 html
+cd build/esp8266
+mv html $BASE/docs/micropython
+cd $BASE
+rm -rf ./mpbuild
 
 # Download terminal programs
 mkdir terminal; cd terminal
@@ -74,10 +98,33 @@ cd ..
 
 # Prepare esp8266 modules to upload
 mkdir micropython; cd micropython
-cp ../micropython-iot-hackathon/example_code/client.py .
+cp $BASE/example_code/client.py .
 cp ../thingflow-python/micropython/*.py .
 cp ../thingflow-python/micropython/sensors/*.py .
-cd ../..
+cd $BASE
+
+echo "Generating README.txt..."
+echo "MicroPython IoT Hackathon Downloads Archive" >README.txt
+echo "===========================================" >>README.txt
+echo "Generated from version $RELEASE on `date`" >>README.txt
+echo "To get started, see docs/micropython-iot-hackathon/index.html" >>README.txt
+echo >>README.txt
+echo "File Layout" >>README.txt
+echo "-----------" >>README.txt
+echo "docs/" >>README.txt
+echo "  micropython-iot-hackathon/" >>README.txt
+echo "  micropython/" >> README.txt
+echo "  thingflow-python/" >>README.txt
+echo "drivers/  (serial drivers for MacOS and Windows)" >>README.txt
+echo "esp8266-20170612-v1.9.1.bin (MicroPython firmware image)" >>README.txt
+echo "example_code/ (from micropython-iot-hackathon repo)" >>README.txt
+echo "micropython/ (ThingFlow and other code for ESP8266)" >>README.txt
+echo "python-tools/ (Python libraries for your laptop)" >>README.txt
+echo "terminal/ (PuTTY for Windows, screen for Linux)" >>README.txt
+echo "thingflow-python/ (ThingFlow source and example code)" >>README.txt
+echo >>README.txt
+echo "Have fun!" >>README.txt
+cd ..
 
 # Create a zip file
 zip -r micropython-iot-software.zip micropython-iot-software
